Abstract tour files.
